pyproject.toml
[project]
name = "hbbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-gigachat>=0.3.12",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "schedule>=1.2.2",
]
